# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Content Review Tool
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  icon: string;
  guidelines: string;
}

export type ReviewIssueType = typeof ReviewIssueType[keyof typeof ReviewIssueType];


export const ReviewIssueType = {
  seo: 'seo',
  google_ads: 'google_ads',
  google_shopping: 'google_shopping',
  gdn: 'gdn',
  content_quality: 'content_quality',
  policy_violation: 'policy_violation',
} as const;

export type ReviewIssueSeverity = typeof ReviewIssueSeverity[keyof typeof ReviewIssueSeverity];


export const ReviewIssueSeverity = {
  critical: 'critical',
  major: 'major',
  minor: 'minor',
  suggestion: 'suggestion',
} as const;

export interface ReviewIssue {
  type: ReviewIssueType;
  severity: ReviewIssueSeverity;
  title: string;
  description: string;
  /** @nullable */
  location: string | null;
  recommendation: string;
}

export interface ReviewResult {
  seoScore: number;
  adsScore: number;
  shoppingScore: number;
  gdnScore: number;
  overallScore: number;
  issues: ReviewIssue[];
  summary: string;
  criticalCount: number;
  majorCount: number;
  minorCount: number;
  suggestionCount: number;
}

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Review {
  id: number;
  url: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  title?: string | null;
  status: ReviewStatus;
  result?: ReviewResult;
  createdAt: string;
  updatedAt: string;
}

export interface CreateReviewBody {
  url: string;
  categoryId: number;
}

export interface AnalyzeContentBody {
  url: string;
  categoryId: number;
  categoryName: string;
  categoryGuidelines: string;
  reviewId: number;
}

export type ReviewStatsReviewsByCategoryItem = {
  categoryName: string;
  count: number;
};

export interface ReviewStats {
  totalReviews: number;
  completedReviews: number;
  /** @nullable */
  avgSeoScore?: number | null;
  /** @nullable */
  avgAdsScore?: number | null;
  /** @nullable */
  avgOverallScore?: number | null;
  reviewsByCategory: ReviewStatsReviewsByCategoryItem[];
}

export type ListReviewsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};