# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Content Review Tool
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reviews
    description: Content review operations
  - name: categories
    description: Product category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of product categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews
      parameters:
        - in: query
          name: categoryId
          schema:
            type: ["integer", "null"]
        - in: query
          name: limit
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a new content review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews/stats:
    get:
      operationId: getReviewStats
      tags: [reviews]
      summary: Get review statistics
      responses:
        "200":
          description: Review statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewStats"

  /reviews/analyze:
    post:
      operationId: analyzeContent
      tags: [reviews]
      summary: Analyze content from a URL using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeContentBody"
      responses:
        "200":
          description: Analysis result (SSE stream)
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        guidelines:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - icon
        - guidelines

    ReviewIssue:
      type: object
      properties:
        type:
          type: string
          enum: [seo, google_ads, google_shopping, gdn, content_quality, policy_violation]
        severity:
          type: string
          enum: [critical, major, minor, suggestion]
        title:
          type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        recommendation:
          type: string
      required:
        - type
        - severity
        - title
        - description
        - location
        - recommendation

    ReviewResult:
      type: object
      properties:
        seoScore:
          type: integer
        adsScore:
          type: integer
        shoppingScore:
          type: integer
        gdnScore:
          type: integer
        overallScore:
          type: integer
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ReviewIssue"
        summary:
          type: string
        criticalCount:
          type: integer
        majorCount:
          type: integer
        minorCount:
          type: integer
        suggestionCount:
          type: integer
      required:
        - seoScore
        - adsScore
        - shoppingScore
        - gdnScore
        - overallScore
        - issues
        - summary
        - criticalCount
        - majorCount
        - minorCount
        - suggestionCount

    Review:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, analyzing, completed, failed]
        result:
          $ref: "#/components/schemas/ReviewResult"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - url
        - categoryId
        - status
        - createdAt
        - updatedAt

    CreateReviewBody:
      type: object
      properties:
        url:
          type: string
        categoryId:
          type: integer
      required:
        - url
        - categoryId

    AnalyzeContentBody:
      type: object
      properties:
        url:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryGuidelines:
          type: string
        reviewId:
          type: integer
      required:
        - url
        - categoryId
        - categoryName
        - categoryGuidelines
        - reviewId

    ReviewStats:
      type: object
      properties:
        totalReviews:
          type: integer
        completedReviews:
          type: integer
        avgSeoScore:
          type: ["number", "null"]
        avgAdsScore:
          type: ["number", "null"]
        avgOverallScore:
          type: ["number", "null"]
        reviewsByCategory:
          type: array
          items:
            type: object
            properties:
              categoryName:
                type: string
              count:
                type: integer
            required:
              - categoryName
              - count
      required:
        - totalReviews
        - completedReviews
        - reviewsByCategory