# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Content Review Tool
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all product categories
 */
export const ListCategoriesResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "icon": zod.string(),
  "guidelines": zod.string()
})
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem)


/**
 * @summary List all reviews
 */
export const ListReviewsQueryParams = zod.object({
  "categoryId": zod.coerce.number().nullish(),
  "limit": zod.coerce.number().nullish()
})

export const ListReviewsResponseItem = zod.object({
  "id": zod.number(),
  "url": zod.string(),
  "categoryId": zod.number(),
  "categoryName": zod.string().nullish(),
  "title": zod.string().nullish(),
  "status": zod.enum(['pending', 'analyzing', 'completed', 'failed']),
  "result": zod.object({
  "seoScore": zod.number(),
  "adsScore": zod.number(),
  "shoppingScore": zod.number(),
  "gdnScore": zod.number(),
  "overallScore": zod.number(),
  "issues": zod.array(zod.object({
  "type": zod.enum(['seo', 'google_ads', 'google_shopping', 'gdn', 'content_quality', 'policy_violation']),
  "severity": zod.enum(['critical', 'major', 'minor', 'suggestion']),
  "title": zod.string(),
  "description": zod.string(),
  "location": zod.string().nullable(),
  "recommendation": zod.string()
})),
  "summary": zod.string(),
  "criticalCount": zod.number(),
  "majorCount": zod.number(),
  "minorCount": zod.number(),
  "suggestionCount": zod.number()
}).optional(),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})
export const ListReviewsResponse = zod.array(ListReviewsResponseItem)


/**
 * @summary Create a new content review
 */
export const CreateReviewBody = zod.object({
  "url": zod.string(),
  "categoryId": zod.number()
})


/**
 * @summary Get a review by ID
 */
export const GetReviewParams = zod.object({
  "id": zod.coerce.number()
})

export const GetReviewResponse = zod.object({
  "id": zod.number(),
  "url": zod.string(),
  "categoryId": zod.number(),
  "categoryName": zod.string().nullish(),
  "title": zod.string().nullish(),
  "status": zod.enum(['pending', 'analyzing', 'completed', 'failed']),
  "result": zod.object({
  "seoScore": zod.number(),
  "adsScore": zod.number(),
  "shoppingScore": zod.number(),
  "gdnScore": zod.number(),
  "overallScore": zod.number(),
  "issues": zod.array(zod.object({
  "type": zod.enum(['seo', 'google_ads', 'google_shopping', 'gdn', 'content_quality', 'policy_violation']),
  "severity": zod.enum(['critical', 'major', 'minor', 'suggestion']),
  "title": zod.string(),
  "description": zod.string(),
  "location": zod.string().nullable(),
  "recommendation": zod.string()
})),
  "summary": zod.string(),
  "criticalCount": zod.number(),
  "majorCount": zod.number(),
  "minorCount": zod.number(),
  "suggestionCount": zod.number()
}).optional(),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})


/**
 * @summary Delete a review
 */
export const DeleteReviewParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Get review statistics
 */
export const GetReviewStatsResponse = zod.object({
  "totalReviews": zod.number(),
  "completedReviews": zod.number(),
  "avgSeoScore": zod.number().nullish(),
  "avgAdsScore": zod.number().nullish(),
  "avgOverallScore": zod.number().nullish(),
  "reviewsByCategory": zod.array(zod.object({
  "categoryName": zod.string(),
  "count": zod.number()
}))
})


/**
 * @summary Analyze content from a URL using AI
 */
export const AnalyzeContentBody = zod.object({
  "url": zod.string(),
  "categoryId": zod.number(),
  "categoryName": zod.string(),
  "categoryGuidelines": zod.string(),
  "reviewId": zod.number()
})