# lib/db/src/schema/reviews.ts
import { pgTable, text, serial, integer, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const categoriesTable = pgTable("categories", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  slug: text("slug").notNull().unique(),
  description: text("description").notNull(),
  icon: text("icon").notNull(),
  guidelines: text("guidelines").notNull(),
});

export const insertCategorySchema = createInsertSchema(categoriesTable).omit({ id: true });
export type InsertCategory = z.infer<typeof insertCategorySchema>;
export type Category = typeof categoriesTable.$inferSelect;

export const reviewStatusEnum = ["pending", "analyzing", "completed", "failed"] as const;

export const reviewsTable = pgTable("reviews", {
  id: serial("id").primaryKey(),
  url: text("url").notNull(),
  categoryId: integer("category_id").notNull().references(() => categoriesTable.id),
  title: text("title"),
  status: text("status", { enum: reviewStatusEnum }).notNull().default("pending"),
  result: jsonb("result"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertReviewSchema = createInsertSchema(reviewsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertReview = z.infer<typeof insertReviewSchema>;
export type Review = typeof reviewsTable.$inferSelect;